Add the toggle action to the sideBar
