+++ src/components/NavBar/NavBar.style.css
@@ -1,0 +1,3 @@
+.navBar-container {
+  z-index: 2;
+}
