+++ src/components/SideBar/SideBar.style.css
@@ -1,6 +1,8 @@
 .side-menu {
+  z-index: 1;
+  list-style: none;
   padding: 1em;
-  width: 11em;
+  width: 14em;
   position: fixed;
   top: 0;
   height: 100vh;
@@ -8,6 +10,33 @@
   background-color: var(--bg-color);
   box-shadow: rgba(129, 129, 129, 0.356) 10px 5px 10px;
 }
+.side-menu >  li {
+  margin-bottom: 1em;
+}
+.side-menu a {
+  text-decoration: none;
+  color: black;
+}
 .side-menu button {
-  
+  border: none;
+}
+.user-container {
+  justify-content: center;
+  flex-direction: column;
+}
+.separator {
+  margin-top: auto;
+  height: 1px;
+  width: 100%;
+  background-color: #e9e9e9;
+}
+.bg-black {
+  background-color: #00000050;
+  position: fixed;
+  top: 0;
+  height: 100vh;
+  width: 100vw;
+}
+.d-none {
+  display: none;
 }
